leaderboard wrapper moved to nampespace.
